ehance: 将 Beat 相关的部分中间运算过程改为使用 double 以确保累计计算精度 (#299)
顺便将 ReadyBeat 改为使用 uint 以简化逻辑

+++ Cyan-Stars/Assets/Scripts/Chart/Beat.cs
@@ -78,14 +78,14 @@ namespace CyanStars.Chart
 
 
         /// <summary>将 Beat 转换为小数表示的拍子</summary>
-        public float ToFloat()
+        public double ToDouble()
         {
             if (Denominator == 0)
             {
                 return IntegerPart;
             }
 
-            return IntegerPart + (float)Numerator / Denominator;
+            return IntegerPart + Numerator / Denominator;
         }
